+++ crates/rust-client/src/lib.rs
@@ -88,7 +88,7 @@ pub mod testing {
 use alloc::sync::Arc;
 
 use miden_objects::crypto::rand::FeltRng;
-use miden_tx::{auth::TransactionAuthenticator, DataStore, TransactionExecutor};
+use miden_tx::{auth::TransactionAuthenticator, DataStore, TransactionExecutor, TransactionProver};
 use rpc::NodeRpcClient;
 use store::{data_store::ClientDataStore, Store};
 use tracing::info;
@@ -113,6 +113,8 @@ pub struct Client<R: FeltRng> {
     /// An instance of [NodeRpcClient] which provides a way for the client to connect to the
     /// Miden node.
     rpc_api: Box<dyn NodeRpcClient + Send>,
+    /// An instance of [TransactionProver] for proving the validity of executed transactions.
+    tx_prover: Arc<dyn TransactionProver>,
     tx_executor: TransactionExecutor,
 }
 
@@ -141,10 +143,11 @@ impl<R: FeltRng> Client<R> {
     ///
     /// Returns an error if the client could not be instantiated.
     pub fn new(
-        api: Box<dyn NodeRpcClient + Send>,
+        rpc_api: Box<dyn NodeRpcClient + Send>,
         rng: R,
         store: Arc<dyn Store>,
         authenticator: Arc<dyn TransactionAuthenticator>,
+        tx_prover: Arc<dyn TransactionProver>,
         in_debug_mode: bool,
     ) -> Self {
         if in_debug_mode {
@@ -156,7 +159,7 @@ impl<R: FeltRng> Client<R> {
         let tx_executor =
             TransactionExecutor::new(data_store, authenticator).with_debug_mode(in_debug_mode);
 
-        Self { store, rng, rpc_api: api, tx_executor }
+        Self { store, rng, rpc_api, tx_prover, tx_executor }
     }
 
     /// Returns a reference to the client's random number generator. This can be used to generate